Trait amethyst_inspector::Inspect

pub trait Inspect<'a>: Component {
type SystemData: SystemData<'a>;
    fn inspect(data: &mut Self::SystemData, entity: Entity) { ... }

    fn can_add(data: &mut Self::SystemData, entity: Entity) -> bool { ... }

    fn can_remove(data: &mut Self::SystemData, entity: Entity) -> bool { ... }

    fn add(data: &mut Self::SystemData, entity: Entity) { ... }

    fn setup(data: &mut Self::SystemData, entity: Option<Entity>) { ... }
}

Any component implementing Inspect and included in your inspect! will show up in the inspector Whether the component is addable is decided by can_add(...)

Associated Types

type SystemData: SystemData<'a>

Loading content...
Provided methods

fn inspect(data: &mut Self::SystemData, entity: Entity)

This method is only ran if the component contains the selected entity

fn can_add(data: &mut Self::SystemData, entity: Entity) -> bool

Decide if this component can be added (e.g. because it requires another component)

fn can_remove(data: &mut Self::SystemData, entity: Entity) -> bool

Decide if this component can be removed (e.g. because it's required by another component)

fn add(data: &mut Self::SystemData, entity: Entity)

fn setup(data: &mut Self::SystemData, entity: Option<Entity>)

This method is ran on all entities, even if none are selected

Loading content...